Add files via upload
Increases the stack size of items by short name. Find the short name of the items on the Internet Rast.
The setting is made in the listed already added names:
// List of items by Shortname
string[] ammoShortNames = {
"ammo.grenadelauncher.buckshot",
"ammo.grenadelauncher.he",
"ammo.grenadelauncher.smoke" and etc.
also the stack size:
SetItemStackSize(shortName, 644);
To add another stack number, I recommend copying this plugin, renaming the function

"AmmoStackPlugin", change the name to your own and change the file name to the same, and change the stack size, as well as list other short names in the listed items.